requirements.txt
aiogram==3.6.0
asyncpg==0.29.0
fastapi==0.111.0
pydantic==2.7.1
python-dotenv==1.0.1
requests==2.31.0
uvicorn==0.29.0
